Refactored angular material imports
